timer and fixed to string

+++ src/main/java/it/polimi/se2018/server/Server.java
@@ -92,7 +92,7 @@ public class Server {
                     }
 
                 }
-            }, SECONDS * 1000);
+            }, (long) SECONDS * 1000);
 
         }
